Moved date utils to separate file.
